+++ derive/src/args/interface.rs
@@ -155,6 +155,10 @@ pub struct InterfaceAttrs {
     pub name: Option<String>,
 
     #[darling(default)]
+    #[darling(rename = "get_type_name")]
+    pub type_name: bool,
+
+    #[darling(default)]
     pub rename_fields: Option<RenameRule>,
 
     #[darling(default)]
@@ -200,6 +204,10 @@ impl FromItemTrait for Interface {
 impl CommonObject for Interface {
     fn get_name(&self) -> Option<&str> {
         self.attrs.name.as_deref()
+    }
+
+    fn should_impl_type_name(&self) -> bool {
+        !self.attrs.type_name
     }
 
     fn get_ident(&self) -> &Ident {