+++ derive/src/args/resolved_object.rs
@@ -27,6 +27,10 @@ pub struct ResolvedObjectAttrs {
     #[darling(default)]
     pub name: Option<String>,
 
+    #[darling(default)]
+    #[darling(rename = "get_type_name")]
+    pub type_name: bool,
+
     #[darling(default, multiple)]
     #[darling(rename = "register")]
     pub registers: Vec<RegisterAttr>,
@@ -48,6 +52,10 @@ from_derive_input!(
 impl CommonObject for ResolvedObject {
     fn get_name(&self) -> Option<&str> {
         self.attrs.name.as_deref()
+    }
+
+    fn should_impl_type_name(&self) -> bool {
+        !self.attrs.type_name
     }
 
     fn get_ident(&self) -> &syn::Ident {