+++ derive/src/args/simple_object.rs
@@ -59,6 +59,10 @@ pub struct SimpleObjectAttrs {
     pub name: Option<String>,
 
     #[darling(default)]
+    #[darling(rename = "get_type_name")]
+    pub type_name: bool,
+
+    #[darling(default)]
     pub rename_fields: Option<RenameRule>,
 
     #[darling(default, multiple)]
@@ -101,6 +105,10 @@ impl CommonInterfaceAttrs for SimpleObject {
 impl CommonObject for SimpleObject {
     fn get_name(&self) -> Option<&str> {
         self.attrs.name.as_deref()
+    }
+
+    fn should_impl_type_name(&self) -> bool {
+        !self.attrs.type_name
     }
 
     fn get_ident(&self) -> &syn::Ident {